requirements.txt
python-aiml==0.9.0